requirements.txt
beautifulsoup4==4.12.2
Flask==2.3.2
requests==2.31.0
segno==1.5.2
gunicorn==21.2.0
python-dotenv==1.0.0
